Switch to `Result`-based error handling
The catch-all error type is a `Box<dyn std::fmt::Display>` because all
we care about is presenting the error to the user.

Moved `did_error` out of the `Shared` type. This makes setting it the
responsibility of `main` rather than handling it in each `apply`
function. This is a much cleaner and simpler approach.